Forms: add helper methods for data updates to CustomFieldHandler
